% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_call_df.R
\name{api_call_df}
\alias{api_call_df}
\title{API Call (dataframe)}
\usage{
api_call_df(start_date_txt, end_date_txt, api_key, root, path, retries = 10)
}
\arguments{
\item{start_date_txt}{A character, indicating the start date for the data required}

\item{end_date_txt}{character, indicating the end date for the data required}

\item{api_key}{a character, the bearer token provided to access the API. Provided by end user.}

\item{root}{A character, as the root path of the API endpoint - usually some kind of http://www.qqq.com/}

\item{path}{A character, as the specific path of the API - i.e. "api/myapi/v1/apione"}

\item{retries}{an optional numeric, indicating how many times the function should try to hit the API in case of an error}
}
\description{
Make sequential calls on an API, one day at a time, and return a concatenated tibble
}
\details{
The only useful external function - this wraps up all the others to sequentially issue
requests against an API one day at a time, since they may overflow if you do more than
that.
}
\examples{
\dontrun{
api_key <- "xxxxx-xxxx-xxxx-xxx-xxxx"
root_path <- "http://www.qqq.com/"
specific_path <- "api/myapi/v1/apione"

start_date <- "2020-06-01"
end_date <- "2020-06-12"

df <- api_call_df(start_date_txt = start_date,
                  end_date_txt = end_date,
                  api_key = api_key,
                  root = root_path,
                  path = specific_path)
}
}
\author{
Matt Simmons mattsimmons@qantas.com.au
}
