% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_call_raw.R
\name{api_call_raw}
\alias{api_call_raw}
\title{API Call (Raw)}
\usage{
api_call_raw(root, path, start_date, end_date, api_key, retries = 10)
}
\arguments{
\item{root}{A character, as the root path of the API endpoint - usually some kind of http://www.qqq.com/}

\item{path}{A character, as the specific path of the API - i.e. "api/myapi/v1/apione"}

\item{start_date}{a character, used in the query as the start of the query period}

\item{end_date}{a character, used in the query as the end of the query period}

\item{api_key}{a character, the bearer token provided to access the API. Provided by end user.}

\item{retries}{an optional numeric, indicating how many times the function should try to hit the API in case of an error}
}
\description{
Calls the API endpoint and returns a list containing the output
}
\details{
This function calls an arbitrary API endpoint and returns the output,
along with some informative error handling. Wouldn't usually be used directly.
}
\author{
Matt Simmons mattsimmons@qantas.com.au
}
