% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_to_tibble.R
\name{element.to.tibble}
\alias{element.to.tibble}
\title{Element to Tibble}
\usage{
element.to.tibble(element)
}
\arguments{
\item{element}{A list, an element of a JSON that could be turned into a tibble.}
}
\description{
This turns an list-element of JSON into a tibble
}
\details{
This function takes an element of a JSON and turns it into a tibble. Not that useful
except as used in other functions
}
\examples{
\dontrun{
out_content_flat <- purrr::flatten(out_content)
all_elements_flat <- purrr::map(out_content_flat, rlist::list.flatten)
all_elements_tibble <- purrr::map(all_elements_flat, element.to.tibble)
}
}
\author{
Matt Simmons mattsimmons@qantas.com.au
}
